// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const CallStatus = {
  PENDING: 'PENDING',
  RINGING: 'RINGING',
  ANSWERED: 'ANSWERED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  BUSY: 'BUSY',
  NO_ANSWER: 'NO_ANSWER',
  CANCELED: 'CANCELED'
} as const

export type CallStatus = (typeof CallStatus)[keyof typeof CallStatus]


export const AmdResult = {
  HUMAN: 'HUMAN',
  MACHINE: 'MACHINE',
  UNDECIDED: 'UNDECIDED',
  TIMEOUT: 'TIMEOUT'
} as const

export type AmdResult = (typeof AmdResult)[keyof typeof AmdResult]